pyproject.toml
[tool.poetry]
name = "youtube-analytics-project"
version = "0.1.0"
description = ""
authors = ["Marat Zaripov <maratos96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
build = "^1.0.3"
isodate = "^0.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
